[unnecessary copy] Introduce new issue type PULSE_READONLY_SHARED_PTR_PARAM
Summary:
This diff introduces a new issue type PULSE_READONLY_SHARED_PTR_PARAM, which is similar to PULSE_CONST_REFABLE in terms of that a parameter passed by-value and it is not modified in the function, but gives a different fix guide:

Consider passing the raw pointer instead of shared pointer unless the callee extends its lifetime.

To (conservatively) check if the lifetime is extended or not in callee, it checks

* if the parameter is captured in lambda, or
* if the parameter is passed to the any functions except `std::shared_ptr`'s observers, e.g. `get`, `operator*`, `operator->`, etc.

Any copy of shared pointer is addressed as lifetime extension, because when copied/assigned, it calls `std::shared_ptr`'s copy/assign constructors and this matches with the second rule above.

Reviewed By: ezgicicek

Differential Revision: D40936801

fbshipit-source-id: d796e94d95f369dda0ee6abcde13acb64b073082