[textual] Support `...` syntax for formals in declarations
Summary:
In Hack we translate the files 1-by-1 and don't know the types of
external functions when we translate them. Still, we somehow need to
declare them and use these declarations at other stages of the
translation.

One possibility is to model such declarations as variadic function, but
that doesn't work well because the calling convention for variadic
functions becomes unclear and generally we conflate variadic functions
with unknown functions.

Instead, we allow the `...` syntax in declarations to mark ProcDecls
with unknown formals. Such ProcDecls are then handled accordingly in
later passes.

Reviewed By: davidpichardie

Differential Revision: D41095070

fbshipit-source-id: 108c5ae3a7e0748e2566957fcc8c222943d6309e