[infer][objc][pulse] Fixing closure-specialize for when a dereference is present
Summary:
Closure specialize fails when there is a mix of fields and dereference accesses as for v6 in the example below.

```
{ roots={ n$0=v6, n$4=v6, &vc=v2 };
       mem  ={ v2 -> { * -> v6 },
               v6 -> { __infer_model_reference_count -> v51,
                       _funcTable -> v7,
                       * -> v8 },
```

f the deref is seen before field accesses then it won't have effect on closure-specialize. But if the deref is seen after seeing the field accesses then it will replace the result by None. Trying closure-specialize fails because it erases all the previous accesses' results when it sees the dereference (*-> v8).

This diff fixes how we treat spurious defers during closure.

Reviewed By: fantazio

Differential Revision: D40674604

fbshipit-source-id: a8d71f25604a21e2cbfa02d98e1884a9e6416b32