[const refable] Suppress issue when parameter is copied to field or global
Summary:
When a passed-by-value parameter is unnecessarily copied to initialize an object field or global variable, a proper fix, in general, is to move the parameter (`std::move`), rather than changing the type of the paramter to const-ref.  This diff suppresses the issues when the parameter is copied to an object field or global.

Note that this checking is intra-procedural.

Reviewed By: ezgicicek

Differential Revision: D41002822

fbshipit-source-id: e3071b61d0b9845489d4bcbb87fe6444d66c75a5