[hack] Make hackc driver work with --force-integration
Summary:
Sometimes it's convenient to use a custom runner for hackc on the
command line. We have a `--force-integration` CLI option but previously
it would always use `hackc` as a binary to run.

Now hack driver checks whether force-integration was used and picks the
program accordingly (e.g. doesn't change the command line at all when
--force-integration hackc is set).

Reviewed By: geralt-encore

Differential Revision: D41218379

fbshipit-source-id: ac6ef4821125d2e5b4a5e7d932aaf1117c6f533d