Upstream bufferoverrun changes (#1692)
Summary:
This pull request contains a part of the patches we wrote with the aim to make BO work on the SIL generated by the Infer Ada frontend (developed at AdaCore).

**Focus on zero false-positive rate**

The main goal was to provide false-positive free analysis (we have different, more heavy-weight analyses that aim at a low or even zero number of false-negatives). Thus, only messages where BO can prove that the problem occurs are reported. The consequence of that is that an imprecision cannot be a cause of a false positive while an unsoundness can. This can cause a clash between our use of BO and the use of BO at Facebook. In order to avoid this clash, we put changes making BO sound and introducing diffs in cost analysis into options.

**New options**
We put the changes introducing undesired diffs - in particular in cost analysis - under options. These options are:

_--bo-sound-unknown-sets-join_: join with an unknown set always result in an unknown set. When disabled, an unknown set behaves as bot.

_--bo-bottom-as-default_: use bottom as a default value instead of unknown. Enabled by default.

_--bo-assume-void_: assume void type as a type of record fields not in type environment. Enabled by default.

_--bo-context-sensitive-allocsites_: Assumes that an allocsite is imported as different allocsite from a callee to a caller in different call contexts. This assumption is ensured by the Ada frontend and makes it possible to do strong updates of locations imported from callees to callers more often. Disabled by default.

**Notation**

Each commit causing diffs in tests output is followed by commits updating test baselines - one commit for each group of tests (c_bufferoverrun, cpp_bufferoverrun, java_bufferoverrun, cpp_performance, and java_performance). Commit messages of commits updating baselines contain an analysis of diffs.

Each commit is marked by one of the following:

[Arrays]: changes in handling of arrays
[Top]: make top default instead of bot
[F]: other fixes
[I]: other improvements

Pull Request resolved: https://github.com/facebook/infer/pull/1692

Reviewed By: ngorogiannis

Differential Revision: D41160322

Pulled By: skcho

fbshipit-source-id: dadfdbad9c9e40183215aff4e5c21188690b896a