[pulse][array] improve access lookup by using canonicalized representation
Summary:
In arrays, indices are represented using abstract values. This means
that if we try to access the same array cell using different abstract
values representing the same concrete value, we will end up with
different cells. This is the problem we are solving here.

This situation was already partially avoided with value unification.
Indeed, if we know that two values are equal then one of them can be
used systematically as the representant for those values.
However, the first value that appears might not always be known as being
equal to the values it represents as shown in `test_intra_change_0_bad`.
A value can therefore take over the representation of the group if it
later finds that it is equal to its members. To continue understanding
what abstract value as index gives what cell, we cannot assume our
previous existing edges to always be canonical. Instead of simply
looking at the cell associated with the given abstract value as index,
we first look for that edge (chances are that this is already the
"canonical edge") and if we cannot find it, then we look for the
canonical representation of the given abstract value in the canonical
representation of the known indices

Reviewed By: dulmarod

Differential Revision: D41083943

fbshipit-source-id: f9dc62cf22197fd7d479c20a0c6fbe0d2a2d9176