[pulse][array] canonicalize arrays at function call
Summary:
When calling a function, the pre of the function may contain information
about cells in an array given as argument. However, these information
were not unified with the ones existing in the calling context,
resulting in potentially two cells with different value information or
attributes when the abstract values used for their indices are actually
known to be equal.

To avoid that mistake and actually propagate cell information from
caller to callee, we do a form of canonicalization where all cell
duplicates are grouped into a canonical one chosen to be the one
accessible from the caller. We choose the one from the caller because
when calling the function the caller knows information about the value
that the callee does not but not the other way around

Reviewed By: dulmarod

Differential Revision: D41083941

fbshipit-source-id: 3cb4c52fd90228d0f3bebc61f8ddf70b02e7b263