[pulse] Incorporate new_eqs in their discovery order
Summary:
New equalities come mostly from union-find operations. Suppose we have
variables x, y, z, and we execute `union(x,y); union(y,z)`. Suppose
that, in this example, union-find picked as class-rep of its result the
second argument of the call to `union`. Then it will also return the
"new equalities" x=y and y=z, in this order. If we treat them as
substitutions x->y and y->z they must be applied in this order; doing
y->z first and then x->y will give us a formula that may still contain
variable y.

To make sure we keep so, we accumulate new equalities, as they are
discovered, in a `RevList` and turn it into a list just before applying
them as substitutions. (One could use just the `List.t` type, of course,
but this way we get some more type safety. The rule is we use
`RevList.t` while discovering equalities, and `List.t` while applying
equalities as substitutions.)

Reviewed By: jvillard

Differential Revision: D40385316

fbshipit-source-id: 8ce02c95c86180a0e2f968301f73b501c56438b8