Unifying Doli and Textual -- 1st part
Summary:
`Doli` and `Textual` should share the same description of the bodies of the functions/procedures.

We will achieve this by making one large parser  with two entries, one for `Textual` and one for `Doli`. These two parsers will share the syntax and parsing for the bodies, and will be using the function `_body`.  And they will be using the same lexer.

*CURRENT STATE* The current Diff is the first part in this journey. Here we prepare `Doli` and `Textual` for the common parser. The current Diff achieves the following

(V1): Unify the tokens used by `Doli` to be the same as those used by `Textual`.

(V2) Remove unnecessary start symbols, and refactor `Doli` tokens into a file. Will do the same for `Textual` later on now but waiting till I merge the parsers.

(V3) Refactor `Textual`, by adding a `body` function, which returns the body.

(V4) Introduce a module `Textual.Body` which is the return type of `_body`. The relation between   `Textual.Body` and`Textual.ProcDesc` still to be determined.

Reviewed By: jvillard

Differential Revision: D41220090

fbshipit-source-id: d14d93326beb4628d8515ba3baab1ce3abc16ce9