[frontend][clang] fix `BinaryConditionalOperator`
Summary:
Using the `?:` operator with opaque exp lead to misordered nodes.
e.g. `y = foo(x) ?: x` lead to instructions similar to
```
if (n$0) {
  y = n$0;
  exit_scope(n$0);
}
else {
  exit_scope(n$0);
  y = x;
}
n$0 = foo(x); // This should appear before the conditions
```

The issue was due to the `n$0 = foo(x)` instruction not belonging to a
node at first while the conditions did. When we group instructions in
`CTrans_utils.collect_controls`, those belonging to nodes are kept in
their nodes and thos not are added to a new node that is placed after
the existing ones.

Now, the `n$0 = foo(x)` instruction belongs to its own
`BinaryConditionalStmtInit` node so when we group all the instructions
in `y = foo(x) ?: x`, the order in which we group them is kept.

Reviewed By: dulmarod

Differential Revision: D41649605

fbshipit-source-id: dd5a866d6ff62ddab8d685f1253b8db33fa64521