[unnecessary copy] Report copies form const & with distinct issue type and informative error message
Summary:
Copies originating from `const &` sources cannot be directly moved. The source type needs to be also modified to remove `const &` as well so that we can move it. This diff adjusts the error message to reflect this suggestion. Developers might prefer not to change the type and hence not to fix the issue without disrupting the whole caller chain so I want to track these issues separately. I suspect these issues will have lower fix rate. To do that, we introduce new const counter parts of two issues:
- unnecessary copy assignment
- unnecessary intermediate copy

Reviewed By: ngorogiannis

Differential Revision: D41804099

fbshipit-source-id: d8f3dcef60f43ef6802b9679b0f6420a26bc96d5