SimpleShape: shape variables captured in a closure
Summary:
When analysing a function, the captured variables were not given a shape in particular. Coupled with
our (non-)translation of exceptions, this could cause an issue when these variables where only used
in an exception handler: they would then indeed been captured (and correctly inferred as such by
Lineage), but would not appear in any instruction of the body of the corresponding procedure, thus
would not be given a shape. This would cause issues when Lineage is trying to do something with these
captured variables and request their shape.

We fix this by first shaping the captured variables before analysing a closure. Thus they will be
present in the summary even if not "used" by the closure.

Reviewed By: rgrig

Differential Revision: D41848205

fbshipit-source-id: 5abb5e7f7e9fcc521335fc6b75b7e60d87e91c14