Redefine the lineage analysis as depending on SimpleShape for the same procedure
Summary:
The previous definition of SimpleLineage used a function that allowed it to depend on SimpleShape
having run for the dependencies of the procedure being analysed, but not for the current one. We
mimic the version that exists for intraprocedural analyses to implement a new function for defining
interprocedural analyses that depend on the summary of another interprocedural analysis for the
procedure that is being analysed, and use it to refactor SimpleLineage.

We still don't actually use the shape result for now, but we do check that it is actually available
and fail otherwise.

Reviewed By: rgrig

Differential Revision: D41883676

fbshipit-source-id: 07974066a260ef8d906d4be63ad7f2206b6fbe4c