[erlang] Check return type against spec
Summary:
For known functions, we used to assume that formals have the type
declared in the function spec. Now we also assert that the returned
value has the type declared in the function spec. This is guarded by an
option (`--erlang-check-return`).

Besides allowing us to find a few more problems, the assertion itself
can be used in Topl properties that care about the type of the value
returned by a function. (For an example, see
`infer/tests/codetoanalyze/erlang/topl/specs/property.topl`.)

Also, this diff refactors the translation of clauses to use a
higher-level API: instead of `Procdesc.Node.t` prefer `ErlangBlock.t`.

Differential Revision: D41622014

fbshipit-source-id: 014bb53c1fdf0f5683538eddb242dcd697fc33ed