Implement field sensitivity for lineage (disabled by default)
Summary:
This is the production-ready implementation of field sensitivity for Lineage, driven by two options,
`--simple-lineage-field-width` and `--simple-lineage-field-depth`. `width` is optional, and if set,
will limit the flat width of a structure to have its fields being distinguished. `depth` is currently
mandatory as the only way to terminate in the presence of cyclic shapes, and sets the maximal number
of subscripted fields from a variable over which any deeper fields will not be distinguished.

Setting either `field-width` or `field-depth` to `0` will effectively make the analysis
field-insensitive, which also remains the default behaviour with `width` unset and `depth` set to
`0`.

Reviewed By: rgrig

Differential Revision: D42007603

fbshipit-source-id: 501e6d9020b159523ea0d098facf30daa3c40e37