[unnecessary copy] Don't report on local copies that are used later
Summary:
We only relied on loaded variables to see if a variable was used later in the code. However, for locally defined variables, this doesn't work for some cases because the frontend doesn't generate loads unlike parameters.

A simple example is:

```
void source_mod_ok() {
  Arr source;
  foo(source);
  source.arr[0] = 9; // Infer translates this as a direct store (no load before) =  *&source.arr[0]:int=9
}
```

This is because Infer handles objects of local variables and parameters differently. When passing an object parameter, the object is constructed at the callsite and then its reference (i.e. pointer) is passed to the function. So, in the function, it dereferences (loads) the parameter when it is being accessed. For a local variable, there is no such dereference required.

This diff addresses this problem by keeping track of the program variables in stores and checks for these for uses.

Reviewed By: skcho

Differential Revision: D42572082

fbshipit-source-id: 62397d40672267d0cf65d66897cf9f1563b6400f