[build] Update menhir to 20220210 (latest)
Summary:
Compared to the previously used version, the new version has a different
code backend. As a result we need either annotate a type of every single
non-terminal or turn on inference and do some manual type
disambiguation.

I chose the latter approach here.

This update is required for ocaml 5.

NOTE: we can't use menhir's new code backend with the Java source parser and flambda because the compilation doesn't terminate so we use the `--table` backend there. Topl and Textual parsers work fine with the new menhir's code backend.

Reviewed By: jvillard

Differential Revision: D42673934

fbshipit-source-id: 8b0b3c77c3c0aaad5c8688f9b9c2596db566abec