[unnecessary copy] Suppress issues on copy ctor/assignment
Summary:
Don't report issues on copy constructor or assignment as they tend to copy objects by design and one cannot simply move as a fix because the source is const-ref.

We could have also checked for the source type but currently there is a bug in frontend that prevents us seeing member field's type as const. Hence, we can't really do that. This is causing huge number of issues in internal fbcode that I would rather suppress immediately.

Reviewed By: skcho

Differential Revision: D42844633

fbshipit-source-id: fe9275d3b03aebe4e3ff6b3e83639b9802cea194