[report][unnecessary copy] Include point of instantiation in traces
Summary:
Issues in templated functions were reported at the template definition with a specialized procname that includes the specialized type. However, it is not clear where these functions are instantiated first, making it difficult to understand some issues like unnecessary copies.

This diff alleviates this problem by including the point of instantiation in the first step of the trace.

Reviewed By: skcho

Differential Revision: D43395985

fbshipit-source-id: 3eda9a9c1db64d235ea5897a39db6bd836a19a91