[textual] Register undefined types used in decls as part of tenv
Summary:
Textual doesn't require that types used inside a module are defined in
the module. This leads to undefined types not being a part of the file's
tenv which in turn causes problems down the line (e.g. PatternMatch not
picking up types when checking againts taint sources/sinks).

This diff adds undefined types used in the module's declarations to the
tenv.

NOTE: the diff doesn't cover types used in alloc and cast expressions.
This will be done separately.

Reviewed By: davidpichardie, ngorogiannis

Differential Revision: D43246895

fbshipit-source-id: f97b4d2e9972ef4e6b7235787fb3a6b61c6e8e7d